khr_get_memory_requirements2 and khr_dedicated_allocation

+++ extensions/khr_bind_memory2/options_bindbuffermemory.go
@@ -15,7 +15,7 @@ import (
 type BindBufferMemoryOptions struct {
 	Buffer       core1_0.Buffer
 	Memory       core1_0.DeviceMemory
-	MemoryOffset uint64
+	MemoryOffset int
 
 	common.HaveNext
 }
